Fixed various compiler warnings

+++ src/GreenWallet/Transfer.cpp
@@ -448,7 +448,7 @@ void transfer(std::shared_ptr<WalletInfo> walletInfo, uint32_t height, bool send
 }
 
 uint64_t calculateNodeFee(uint64_t amount) {
-	uint32_t node_fee = static_cast<int64_t>(amount * 0.0025);
+	uint64_t node_fee = static_cast<int64_t>(amount * 0.0025);
 	if (node_fee > (uint64_t)10000000000000)
 		node_fee = (uint64_t)10000000000000;
 	return node_fee;
